Tiles should take les memory now.

+++ include/tile.hpp
@@ -14,8 +14,8 @@ class Tile {
         Tile();
         ~Tile();
 
-        Tile(const std::string in_name, const Movement_cost in_cost, unsigned int * in_stats[4]);
-        Tile(const std::string in_name, const Movement_cost in_cost, unsigned int * in_stats[4], const bool in_inside);
+        Tile(const std::string in_name, const Movement_cost in_cost, short int * in_stats[4]);
+        Tile(const std::string in_name, const Movement_cost in_cost, short int * in_stats[4], const bool in_inside);
 
         Movement_cost getCoststruct();
         std::unordered_map<std::string, unsigned char> getCostmap();
